added clippi skip

+++ crates/ffmpeg/src/scaler.rs
@@ -20,6 +20,8 @@ unsafe impl Send for Scaler {}
 
 impl Scaler {
     /// Creates a new `Scaler` instance.
+    /// The unnecessary cast is needed for Scaler to work on windows.
+    #[allow(clippy::unnecessary_cast)]
     pub fn new(
         input_width: i32,
         input_height: i32,
@@ -37,7 +39,7 @@ impl Scaler {
                 width,
                 height,
                 pixel_format,
-                SWS_BILINEAR,
+                SWS_BILINEAR as i32,
                 std::ptr::null_mut(),
                 std::ptr::null_mut(),
                 std::ptr::null(),
